Keep the component in addon and reexport for resolver.
